Added cast for 64-bit machines
git-svn-id: file://localhost/var/lib/svn/m6809@3 e160def4-4833-0410-8f62-e3393c40948d

+++ sim/miscutils.c
@@ -28,7 +28,8 @@ void *mmalloc(size_t n)
 
   p = (void *)malloc(n);
   if (!p) {
-    fprintf(stderr, "Not enough memory to allocate %d bytes\n", n);
+    /* Note cast to make code portable to 64-bit systems */
+    fprintf(stderr, "Not enough memory to allocate %ld bytes\n", (long)n);
     abort();
   }
   return p;
